requirements.txt
python-docx
numpy
seaborn
matplotlib
scipy
pandas
itertools